This complies, at least
