+++ src/schemas/number.rs
@@ -1,0 +1,19 @@
+use std::fmt;
+
+use crate::Number;
+
+/// A number schema
+#[derive(Debug, PartialEq)]
+pub struct NumberSchema {
+    pub minimum: Option<Number>,
+    pub maximum: Option<Number>,
+    pub exclusive_minimum: Option<Number>,
+    pub exclusive_maximum: Option<Number>,
+    pub multiple_of: Option<Number>,
+}
+
+impl fmt::Display for NumberSchema {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "Number {:?}", self)
+    }
+}
