+++ src/schemas/one_of.rs
@@ -1,4 +1,3 @@
-use serde::{Deserialize, Serialize};
 /// The `oneOf` schema is a schema that matches if any of the schemas in the `oneOf` array match.
 /// The schemas are tried in order, and the first match is used. If no match is found, an error is added
 /// to the context.
@@ -9,8 +8,7 @@ use crate::{format_vec, YamlSchema};
 /// The `oneOf` schema is a schema that matches if any of the schemas in the `oneOf` array match.
 /// The schemas are tried in order, and the first match is used. If no match is found, an error is added
 /// to the context.
-#[derive(Debug, Default, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "camelCase")]
+#[derive(Debug, Default, PartialEq)]
 pub struct OneOfSchema {
     pub one_of: Vec<YamlSchema>,
 }
@@ -21,37 +19,10 @@ impl fmt::Display for OneOfSchema {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::engine::Context;
-    use crate::validation::Validator;
-    use crate::{TypeValue, TypedSchema, YamlSchema, YamlSchemaNumber};
-
-    #[test]
-    fn test_one_of_schema() {
-        let schemas = vec![
-            YamlSchema::TypedSchema(Box::new(TypedSchema {
-                r#type: TypeValue::number(),
-                multiple_of: Some(YamlSchemaNumber::Integer(5)),
-                ..Default::default()
-            })),
-            YamlSchema::TypedSchema(Box::new(TypedSchema {
-                r#type: TypeValue::number(),
-                multiple_of: Some(YamlSchemaNumber::Integer(3)),
-                ..Default::default()
-            })),
-        ];
-
-        let schema = OneOfSchema { one_of: schemas };
-        println!("{}", schema);
-        let root_schema = YamlSchema::OneOf(schema);
-        let context = Context::new(false);
-        assert!(root_schema
-            .validate(
-                &context,
-                &serde_yaml::Value::Number(serde_yaml::Number::from(5.0))
-            )
-            .is_ok());
+impl From<crate::deser::OneOfSchema> for OneOfSchema {
+    fn from(deserialized: crate::deser::OneOfSchema) -> Self {
+        OneOfSchema {
+            one_of: deserialized.one_of.into_iter().map(|s| s.into()).collect(),
+        }
     }
 }
