+++ src/validation/objects.rs
@@ -3,8 +3,9 @@ use log::debug;
 use std::collections::HashMap;
 
 use super::Validator;
+use crate::schemas::{BoolOrTypedSchema, TypedSchema};
 use crate::validation::Context;
-use crate::{AdditionalProperties, TypeValue, TypedSchema, YamlSchema, YamlSchemaError};
+use crate::{YamlSchema, YamlSchemaError};
 
 pub fn try_validate_value_against_properties(
     context: &Context,
@@ -31,51 +32,20 @@ pub fn try_validate_value_against_additional_properties(
     context: &Context,
     key: &String,
     value: &serde_yaml::Value,
-    additional_properties: &AdditionalProperties,
+    additional_properties: &BoolOrTypedSchema,
 ) -> Result<bool, YamlSchemaError> {
     let sub_context = context.append_path(key);
 
     match additional_properties {
         // if additional_properties: true, then any additional properties are allowed
-        AdditionalProperties::Boolean(true) => { /* noop */ }
+        BoolOrTypedSchema::Boolean(true) => { /* noop */ }
         // if additional_properties: false, then no additional properties are allowed
-        AdditionalProperties::Boolean(false) => {
+        BoolOrTypedSchema::Boolean(false) => {
             context.add_error(format!("Additional property '{}' is not allowed!", key));
             // returning `false` signals fail fast
             return Ok(false);
         }
-        // if additional_properties: { type: <string> } or { type: [<string>] }
-        // then we validate the additional property against the type schema
-        AdditionalProperties::Type { r#type } => {
-            // get the list of allowed types
-            let allowed_types = r#type.as_list_of_allowed_types();
-            debug!(
-                "validate_object_mapping: allowed_types: {}",
-                allowed_types.join(", ")
-            );
-            // TODO: Check if the value _is_ valid for any of the allow types
-            // return Ok if so
-            // return an error otherwise
-            // check if the value is _NOT_ valid for any of the allowed types
-            let allowed = allowed_types.iter().all(|allowed_type| {
-                let sub_schema = TypedSchema {
-                    r#type: TypeValue::from_string(allowed_type.clone()),
-                    ..Default::default()
-                };
-                debug!(
-                    "Validating additional property '{}' with schema: {:?}",
-                    key, sub_schema
-                );
-                sub_schema.validate(&sub_context, value).is_ok()
-            }); // if the value is not valid for any of the allowed types, then we return an error immediately
-            if !allowed {
-                context.add_error(format!(
-                    "Additional property '{}' is not allowed. No allowed types matched!",
-                    key
-                ));
-                return Ok(false);
-            }
-        }
+        _ => return not_yet_implemented!(),
     }
     Ok(true)
 }
